Move PaperPlotter class to its own file
